refactor: 패키지 구조 변경 및, 관련 test 삭제
