+++ src/main/java/greedy/greedybot/domain/message/ScheduledMessageRepository.java
@@ -1,7 +1,5 @@
 package greedy.greedybot.domain.message;
 
-import greedy.greedybot.application.message.dto.ScheduledMessage;
-
 import java.util.List;
 import java.util.Optional;
 
@@ -11,7 +9,7 @@ public interface ScheduledMessageRepository {
 
     void deleteScheduledMessage(final String formId);
 
-    Optional<ScheduledMessage> findByFormId(final String formId);
+    Optional<ScheduledMessage> findById(final String formId);
 
     List<ScheduledMessage> findAll();
 }
